pyproject.toml
[tool.poetry]
name = "fastapi-tortoise-api-base"
version = "0.1.0"
description = ""
authors = ["plusiv"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
fastapi = "^0.103.0"
tortoise-orm = {extras = ["asyncmy"], version = "^0.20.0"}
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
aiohttp = "^3.8.5"
aiodns = "^3.0.0"
email-validator = "^2.0.0.post2"
casbin = "^1.24.0"
pydantic-settings = "^2.0.3"


[tool.poetry.group.dev.dependencies]
tortoise-cli = "^0.1.2"
faker = "^19.3.1"
aerich = {git="https://github.com/tortoise/aerich.git", branch="dev"}
black = "^23.7.0"
pre-commit = "^3.3.3"
httpx = "^0.24.1"
pytest = "^7.4.0"
asgi-lifespan = "^2.1.0"
ruff = "^0.0.285"
mypy = "^1.5.1"

[tool.ruff]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "I",   # isort
  "PL",  # pylint

]
ignore = ["E501"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

target-version = "py311"

[tool.aerich]
tortoise_orm = "app.core.settings.TORTOISE_ORM"
location = "app/database/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
